Add files via upload
I submitted the cookie sim. late because it is my first time using GitHub. I thought I had submitted it, but did not, and I had not realized this because I am still getting used to this website. I also thought that I was working with a group and my classmate had turned it in but it turns out I did not join the repository.